@@ -8,11 +8,11 @@
   <version>1.0-SNAPSHOT</version>
 
   <properties>
-    <maven.compiler.release>11</maven.compiler.release>
+    <maven.compiler.release>17</maven.compiler.release>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
     <version.org.optaplanner>8.33.0-SNAPSHOT</version.org.optaplanner>
-    <version.org.springframework.boot>2.7.4</version.org.springframework.boot>
+    <version.org.springframework.boot>3.0.1</version.org.springframework.boot>
 
     <version.compiler.plugin>3.8.1</version.compiler.plugin>
     <version.surefire.plugin>3.0.0-M5</version.surefire.plugin>
